refactoring and comments
