+++ app/src/main/java/dev/leonlatsch/photok/model/database/PhotokDatabase.kt
@@ -21,14 +21,15 @@ import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
 import dev.leonlatsch.photok.model.database.PhotokDatabase.Companion.VERSION
 import dev.leonlatsch.photok.model.database.dao.PasswordDao
-import dev.leonlatsch.photok.model.database.entity.Photo
 import dev.leonlatsch.photok.model.database.dao.PhotoDao
 import dev.leonlatsch.photok.model.database.entity.Password
+import dev.leonlatsch.photok.model.database.entity.Photo
 
 /**
  * Abstract Room Database.
  *
  * @since 1.0.0
+ * @author Leon Latsch
  */
 @Database(
     entities = [
