+++ app/src/main/java/dev/leonlatsch/photok/model/database/dao/PhotoDao.kt
@@ -24,6 +24,7 @@ import dev.leonlatsch.photok.model.database.entity.Photo
  * Data Access Object for [Photo] Entity.
  *
  * @since 1.0.0
+ * @author Leon Latsch
  */
 @Dao
 interface PhotoDao {
