+++ app/src/main/java/dev/leonlatsch/photok/ui/settings/SettingsFragment.kt
@@ -20,6 +20,12 @@ import android.os.Bundle
 import androidx.preference.PreferenceFragmentCompat
 import dev.leonlatsch.photok.R
 
+/**
+ * Preference Fragment. Loads preferences from xml resource.
+ *
+ * @since 1.0.0
+ * @author Leon Latsch
+ */
 class SettingsFragment : PreferenceFragmentCompat() {
     override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
         setPreferencesFromResource(R.xml.settings, rootKey)
